move sulten views
