+++ backend/samfundet/urls.py
@@ -8,8 +8,8 @@ from rest_framework import routers
 from django.urls import path, include
 
 import samfundet.view.event_views
+import samfundet.view.sulten_views
 import samfundet.view.general_views
-import samfundet.sulten_views.sulten_views
 
 from . import views
 from .view import recruitment_views
@@ -30,12 +30,12 @@ router.register('user-preference', views.UserPreferenceView, 'user_preference')
 router.register('saksdokument', samfundet.view.general_views.SaksdokumentView, 'saksdokument')
 router.register('profile', views.ProfileView, 'profile')
 router.register('permissions', views.PermissionView, 'permissions')
-router.register('menu', samfundet.sulten_views.sulten_views.MenuView, 'menu')
-router.register('menu-items', samfundet.sulten_views.sulten_views.MenuItemView, 'menu_items')
-router.register('food-preference', samfundet.sulten_views.sulten_views.FoodPreferenceView, 'food_preference')
-router.register('food-category', samfundet.sulten_views.sulten_views.FoodCategoryView, 'food_category')
-router.register('booking', samfundet.sulten_views.sulten_views.BookingView, 'booking')
-router.register('table', samfundet.sulten_views.sulten_views.TableView, 'table')
+router.register('menu', samfundet.view.sulten_views.MenuView, 'menu')
+router.register('menu-items', samfundet.view.sulten_views.MenuItemView, 'menu_items')
+router.register('food-preference', samfundet.view.sulten_views.FoodPreferenceView, 'food_preference')
+router.register('food-category', samfundet.view.sulten_views.FoodCategoryView, 'food_category')
+router.register('booking', samfundet.view.sulten_views.BookingView, 'booking')
+router.register('table', samfundet.view.sulten_views.TableView, 'table')
 router.register('textitem', samfundet.view.general_views.TextItemView, 'text_item')
 router.register('interview-rooms', views.InterviewRoomView, 'interview_rooms')
 router.register('infobox', samfundet.view.general_views.InfoboxView, 'infobox')
@@ -57,7 +57,7 @@ router.register('recruitment-applications-for-position', views.RecruitmentApplic
 router.register('interview', views.InterviewView, 'interview')
 
 ######## Lyche #########
-router.register('create-reservation', samfundet.sulten_views.sulten_views.ReservationCreateView, 'create_reservation')
+router.register('create-reservation', samfundet.view.sulten_views.ReservationCreateView, 'create_reservation')
 
 app_name = 'samfundet'
 
@@ -84,7 +84,7 @@ urlpatterns = [
     path('webhook/', views.WebhookView.as_view(), name='webhook'),
     path('gangtypes/<int:organization>/', samfundet.view.general_views.GangTypeOrganizationView.as_view(), name='gangsorganized'),
     ########## Lyche ##########
-    path('check-reservation/', samfundet.sulten_views.sulten_views.ReservationCheckAvailabilityView.as_view(), name='check_reservation'),
+    path('check-reservation/', samfundet.view.sulten_views.ReservationCheckAvailabilityView.as_view(), name='check_reservation'),
     ########## Recruitment ##########
     path('active-recruitments/', views.ActiveRecruitmentsView.as_view(), name='active_recruitments'),
     path('recruitment-positions/', views.RecruitmentPositionsPerRecruitmentView.as_view(), name='recruitment_positions'),
