+++ backend/samfundet/view/sulten_views.py
